Sites API endpoint done

+++ backend/src/controllers/sites.ts
@@ -4,34 +4,19 @@ const { Site, Subscription, Country, Segment, SiteSegment } = require("../sequel
 
 
 export const getAllSites = async (req: Request, res: Response) => {
-
-  // TODO: Validate req: user
-
-  // FROM Klajdi
-  // const session = req.session;
-
-  // req.session.user = {
-  //   id: req.sessionID,
-  // };
-
-  // console.log({ session, sessionId: req.sessionID });
   console.log('** Running controller: getAllSites');
-  // return res.status(200).json({success: 'You have reached getAllSites controller'});
   try {
+
     // const sites = await Site.findAll({
-    //   include: [
-    //     {
-    //       model: SiteSegment,
-    //       include: [Site, Segment]
-    //     }
-    //   ]
+    //   include: [Subscription, Country,
+    //     {model: SiteSegment, include: Segment}
+    //   ],
+    //   attributes: ['name'],
     // });
-    // const sites = await Site.findAll({include: [Country, Subscription]});
-    const sites = await Site.findAll({include: Segment});
-    // const segments = await Segment.findAll({include: Site});
-    // console.log('*** sites variable: ', segments);
-    // console.log('*** sites variable: ', sites);
-    // return res.status(200).json(segments);
+    // const sites = await SiteSegment.findAll({ include: { model: Segment, include: Site} });
+    // const sites = await SiteSegment.findAll({ include: Site });
+    const sites = await Site.findAll();
+
     return res.status(200).json(sites);
   } catch (error) {
     return res.status(500).json({error: 'Server error - could not find release toggles...'});
@@ -42,12 +27,14 @@ export const getOneSite = async (req: Request, res: Response) => {
   console.log('** Running controller: getOneReleaseToggle');
 
   // TODO: Validate req.params.id
-  // Make sure to check if it is an INT / NUMBER
+  // Check if req.params.id is a number
+  if (!Number(req.params.id)) { return res.json({error: 'ID is not a number'}); }
 
   try {
-    const site = await Site.findByPk(req.params.id, {include: [Country, Subscription, Segment]});
+    const site = await Site.findByPk(req.params.id, {include: [Country, Subscription]});
     // const site = await Site.findByPk(req.params.id, {include: [Country as 'country', Subscription as 'sub']});
 
+    // Check if there is data with the provided ID
     if (site === null) { return res.status(404).json({message: 'This release toggle does not exist...'}); }
 
     return res.status(200).json(site);
@@ -65,17 +52,16 @@ export const createSite = async (req: Request, res: Response) => {
 
   // TODO: Validate Data
   if (!req.body.name) { return res.status(400).json({error: 'Name must be defined..'}); }
-  if (!req.body.description) { return res.status(400).json({error: 'Description must be defined..'}); }
-  if (!req.body.userId) { return res.status(400).json({error: 'User ID must be defined..'}); }
+  if (!req.body.countryId) { return res.status(400).json({error: 'countryId must be defined..'}); }
+  if (!req.body.subscriptionId) { return res.status(400).json({error: 'subscriptionId must be defined..'}); }
 
   try {
     const site = await Site.create({
-      // name: req.body.name,
-      // description: req.body.description,
-      // releaseAt: new Date(),
-      // createdAt: new Date(),
-      // updatedAt: new Date(),
-      // userId: req.body.userId,
+      name: req.body.name,
+      createdAt: new Date(),
+      updatedAt: new Date(),
+      countryId: req.body.countryId,
+      subscriptionId: req.body.subscriptionId,
     });
     return res.status(201).json(site);
   } catch (error) {
@@ -92,13 +78,15 @@ export const updateOneSite = async (req: Request, res: Response) => {
 
   interface Payload {
     name?: string;
-    description?: string;
-    userId?: number;
+    countryId?: string;
+    subscriptionId?: number;
   }
   let payload: Payload = {}
-  // if (req.body.name) { payload['name'] = req.body.name }
-  // if (req.body.description) { payload['description'] = req.body.description }
-  // if (req.body.userId) { payload['userId'] = req.body.userId }
+  if (req.body.name) { payload['name'] = req.body.name }
+  if (req.body.countryId) { payload['countryId'] = req.body.countryId }
+  if (req.body.subscriptionId) { payload['subscriptionId'] = req.body.subscriptionId }
+
+  // return res.json(payload);
 
   try {
     const site = await Site.update(payload, {
@@ -119,6 +107,9 @@ export const deleteOneSite = async (req: Request, res: Response) => {
   console.log('id: ',req.params.id, 'type of id: ', typeof req.params.id);
 
   // TODO: Validate req.params.id
+  if (!Number(req.params.id)) { return res.json({error: 'ID is not a number'}); }
+
+  // return res.json({delete: 'Are you sure?'});
 
   try {
     const site = await Site.destroy({
